Create a new pull request by comparing changes

+++ DTO/OrderDto.cs
@@ -2,6 +2,6 @@ namespace BookStoreManage.DTO;
 
 public class OrderDto
 {
-    public double OrderStatus { get; set; }
+    public int OrderStatus { get; set; }
     public int AccountID { get; set; }
 }
